+++ src/conf/serverconf.hpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 17:15:18 by emadriga          #+#    #+#             */
-/*   Updated: 2023/05/16 17:30:39 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/20 20:36:01 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,9 @@ namespace ft
 /**
  * @param address	(std::string)
  * @param port		(unsigned short)
- * @param root		(std::string)
+ * @param default_root		(std::string)
  * @param locations	(std::map<std::string, location>)
+ * @param filetypes (const Filetypes &)
  * @warning Declare an object with ft::serverconf.
 */
 class serverconf{
@@ -39,16 +40,13 @@ class serverconf{
 		const Filetypes &					filetypes;
 
 		//Constructor
-		serverconf(const Filetypes & types)
-			:	filetypes(types)
-		{};
+		serverconf(const Filetypes & types);
 
 		//Destructor
-		~serverconf()
-		{
-			if(locations.size())
-				locations.clear();
-		};
+		~serverconf();
+
+		//Copy constructor
+		serverconf(const ft::serverconf & tocopy);
 
 };