+++ src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 17:15:45 by emadriga          #+#    #+#             */
-/*   Updated: 2023/05/20 17:29:36 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/20 20:48:37 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,7 +101,7 @@ int main(int argc, char **argv)
 			getfilename = "test_files/get_request_dir";
 		else
 			getfilename = "test_files/get_request";
-		Request request(file_reader(getfilename, &read_status), newconf.servers[0]);
+		const Request request(file_reader(getfilename, &read_status), newconf.servers[0]);
 		if (read_status)
 			std::cout << TXT_COLOR_RED << "SOMETHING WENT WRONG IN THE MAIN!" << TXT_RESET << std::endl << "Error code: " << read_status << std::endl;
 		Response response(request);