+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-05-20 - Jorge
+### Added
+- Redirection support (needs testing).
+- `ft::serverconf` copy constructor.
+
 ## 2023-05-20 - Edu
 ### Added
 - Added `exceptions` at loading `conf`, added scripts to enable/disable this exceptions on /src/conf